bugfix: fix private variable access in ConfigClass causing compile
Despite the GetClass method returning an instance of ConfigClass we
can't access a private variable from outside the class instance.

The diag build of the game refused to compile, not certain whether
it is an issue on the retail version but we can fix it by using
the appropriate setters anyway.